CPPprog6 initial commit
initialize 3 floats (x,y,z). Read in values for x,y,z. Initialize and define 'rarea' & 'sarea' integer values as (x*y) and (z*z), respectively. 

If x,y AND z are greater than or equal to .1, check if z is less than x and y. If so, print "Ja Ja'. Othwerwise print "Nein Nein".

If x,y AND z are greater than or equal to .1, check if rarea is greater than sarea. If so print "Rectangle". If rarea is less than sarea, print "Square". 

If x,y AND z are greater than or equal to .1, and rarea>sarea, rarea<area conditions are both false, print "Equal"

If x,y AND z are NOT greater than or equal to .1, do not check any other conditions and print "Invalid Side Values"